backend/requirements.txt
Django
django-cors-headers
django-rest-framework
psycopg2-binary
gunicorn
pillow
gpxpy
geopy
haversine